Refactor equation systems declarations
