+++ src/equation_systems/PDETraits.H
@@ -24,29 +24,6 @@ struct ScalarTransport
     static constexpr bool multiply_rho = true;
 };
 
-struct ICNS : VectorTransport
-{
-    using MLDiffOp = amrex::MLTensorOp;
-    using SrcTerm = MomentumSource;
-
-    static std::string pde_name() { return "ICNS"; }
-    static std::string var_name() { return "velocity"; }
-};
-
-struct Density : ScalarTransport
-{
-    static std::string pde_name() { return "Density"; }
-    static std::string var_name() { return "density"; }
-
-    static constexpr bool multiply_rho = false;
-};
-
-struct Temperature : ScalarTransport
-{
-    static std::string pde_name() { return "Temperature"; }
-    static std::string var_name() { return "temperature"; }
-};
-
 struct TKE : ScalarTransport
 {
     static std::string pde_name() { return "TKE"; }
